Updated to allow standalone launching
Can now launch as a standalone app by removing LaunchActivity from the
manifest and specifying a default app in Configuration
Configuration now uses "Glyde" instead of "FrontEnd"
