+++ app/src/main/AndroidManifest.xml
@@ -7,6 +7,7 @@
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.INTERNET" />
 
+    <!-- application "name" will use that class as startup instead of an activity -->
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -15,11 +16,21 @@
         <activity
             android:name="com.cylexia.mobile.glyde.ViewActivity"
             android:label="@string/app_name" >
-            <!--<intent-filter>
-                 <action android:name="android.intent.action.MAIN" />
+            <!--
+                -v- CUSTOMISE -v-
+                Add the below "intent-filter" block and remove the whole "LaunchActivity"
+                "activity" section to launch a standalone script
+            -->
+            <!--
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>-->
+            </intent-filter>
+            -->
+            <!--
+                -^- CUSTOMISE -^-
+            -->
         </activity>
         <activity
             android:name="com.cylexia.mobile.glyde.LaunchActivity"
