+++ app/src/main/assets/hashcode.app
@@ -1,5 +1,3 @@
 title=HashCode Generator;
 icon=#ic_launcher.png;
 script=#hashcode.script;
-
-        if( FrontEnd._buttons_last > -1 ) then
