+++ app/src/main/java/com/cylexia/mobile/glyde/Configuration.java
@@ -4,6 +4,11 @@ package com.cylexia.mobile.glyde;
  * Created by sparx104 on 30/04/2015.
  */
 public interface Configuration {
-	public final static String ROOT_PATH = "FrontEnd";
+	// -v- CUSTOMISE -v-
+	// The path to read & write files to in the storage
+	public final static String ROOT_PATH = "Glyde";
 
+	// The standalone app file if using without LaunchActivity
+	public final static String STANDALONE_APP_FILE = "#default.app";
+	// -^- CUSTOMSE -^-//
 }
